+++ dydapsconfigurablepacks.php
@@ -60,7 +60,7 @@ final class DydapsConfigurablePacks extends Module
     {
         $this->name = 'dydapsconfigurablepacks';
         $this->tab = 'administration';
-        $this->version = '1.4.0';
+        $this->version = '1.4.1';
         $this->author = 'DYDAPS';
         $this->need_instance = 0;
         $this->bootstrap = true;